% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{breast}
\alias{breast}
\title{Dados de cancer de mama}
\format{Um  \code{data.frame} com 94 linhas e 4 variaveis.
\describe{
   \item{left}{}
   \item{right}{}
   \item{ther}{}
   \item{cens}{indicador de censura (0 se censura)}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(breast)
}
\description{
Estudo para avaliar o efeito de diferentes tratamentos contra o cancer de mama.
A variavel de interesse e o tempo ate se observar uma retracao severa da mama
}
\examples{
data(breast)


}
\references{
Klein e Moeschberger (1997)
}
\keyword{datasets}
