% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{cattle}
\alias{cattle}
\title{Estudo raca nelore}
\format{Um  \code{data.frame} com 155 linhas e 6 variaveis.
\describe{
   \item{id}{identificador do animal}
   \item{sire}{}
   \item{tempo}{}
   \item{censura}{indicador de censura (0 se censura)}
   \item{sex}{sexo do animal}
   \item{agedam}{}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(cattle)
}
\description{
Dado com o tempo de crecimento de gado da raca nelore
}
\examples{
data(cattle)


}
\keyword{datasets}
