% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{desmame}
\alias{desmame}
\title{Dados de aleitamento materno}
\format{Um  \code{data.frame} com 150 linhas e 13 variaveis.
\describe{
   \item{tempo}{empo de aleitamento materno em meses}
   \item{cens}{indicador de censura (0 indica censura)}
   \item{id}{identificacao da mae}
   \item{V1}{experiencia com amamentacao (1 se nao)}
   \item{V2}{numero de filhos vivos (1 se >2}
   \item{V3}{conceito materno sobre tempo ideal de amamentacao (1 se <= 6 meses)}
   \item{V4}{dificuldade para amamentar nos primeiros dias pos-parto (1 se sim)}
   \item{V5}{tipo de servico em que realizou o pre-natal (1 se privado ou convenio, 0 se publico)}
   \item{V6}{recebeu exclusivamente leite materno na maternidade (1 se nao)}
   \item{V7}{a crianca tem contato com o pai (1 se nao)}
   \item{V8}{renda per capita em salario minimo/mes (1 se < 1 SM)}
   \item{V9}{peso ao nascimento (1 se <2.5kg)}
   \item{V10}{tempo de separacao mae-filho pos-parto (1 se > 6)}
   \item{V11}{permanencia no bercario (1 se sim)}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(desmame)
}
\description{
Dados com tempo de aleitamento materno e caracteristicas demograficas e comportamentais.
}
\examples{
data(desmame)
}
\keyword{datasets}
