% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{hepatite}
\alias{hepatite}
\title{Efeito de esteroides no tratamento de hepatite}
\format{Um  \code{data.frame} com 29 observacoe e variaveis.
\describe{
   \item{tempos}{tempo de vida em semanas}
   \item{cens}{indicador de censura (0 indica censura)}
   \item{grupos}{grupo de tratamento (controle, esteroide)}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(hepatite)
}
\description{
Dados de sobrevivencia de um experimento aleatorizado com efeito de esteroides no tempo de vidas de pacientes com hepatite.
}
\examples{
data(hepatite)
}
\references{
Gregoryet al.,1975.
}
\keyword{datasets}
