% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{hg2}
\alias{hg2}
\title{Hormonio de Crescimento}
\format{Um  \code{data.frame} com 80 linhas e 7 variaveis.
\describe{
   \item{tempos}{}
   \item{cens}{indicador de censura (0 se centura)}
   \item{raca}{}
   \item{ialtura}{}
   \item{trauma}{}
   \item{recemnas}{}
   \item{renda}{}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(hg2)
}
\description{
add description
}
\examples{
data(hg2)


}
\keyword{datasets}
