% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{laringe}
\alias{laringe}
\title{Dados estudo sobre cancer de laringe}
\format{Um  \code{data.frame} com 210 linhas e 8 variaveis.
\describe{
   \item{id}{identificacao do paciente}
   \item{tempos}{tempo ate a morte do paciente (meses)}
   \item{cens}{indicador de censura (0 se censura)}
   \item{idade}{idade em anos}
   \item{estagio}{estagio da doenca (1 se tumor primario, 2 se envolve nodulo, 3 se metastase e 4 combinacao dos estagios anteriores)}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(mang)
}
\description{
Dados com o tempo de vida de pacientes com cancer de laringe em diferentes estagios.
}
\examples{
data(mang)


}
\references{
Klein e Moeschberger(1997)
}
\keyword{datasets}
