% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{leucemia_cat}
\alias{leucemia_cat}
\title{Dados com tempo de vida de pacientes com leucemia}
\format{Um  \code{data.frame} com 103 linhas e 10 variaveis.
\describe{
   \item{leuini}{quantidade de leucocitos em 1000 leucocitos por mm^3}
   \item{tempos}{tempo de vida a partir remissao ate a recidiva ou morte}
   \item{cens}{indicador de censura (0 indica censura)}
   \item{idade}{idade em meses}
   \item{zpeso}{peso padronizado pela idade e sexo}
   \item{zest}{altura padronizada pela idade e sexo}
   \item{pas}{porcentagem de linfoblastos}
   \item{vac}{porcentagem de vacuolos}
   \item{risk}{fator de risco em \%}
   \item{r6}{indicador de sucesso da remissao (1 indica sucesso)}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(leucemia_cat)
}
\description{
Dados com tempo de vida e outras caracteristicas (binarizadas) de criancas com leucemia (LLA) de Belo Horizonte, Brasil.
}
\examples{
data(leucemia_cat)
}
\references{
Vianaet al.(1994) e Colosimoet al.(1992).
}
\keyword{datasets}
