% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{mang}
\alias{mang}
\title{Dados de tempo de vida de mangueiras}
\format{Um  \code{data.frame} com 210 linhas e 8 variaveis.
\describe{
   \item{ano}{}
   \item{ti}{}
   \item{cens}{}
   \item{li}{}
   \item{ui}{}
   \item{copa}{}
   \item{cavalo}{}
   \item{bloco}{}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(mang)
}
\description{
Dados com a resistencia a pragas de mangueiras e com censura intervalar.
}
\examples{
data(mang)
}
\references{
Chalitaet al.(1999)
}
\keyword{datasets}
