% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{ovario}
\alias{ovario}
\title{Dados cancer de ovario}
\format{Um  \code{data.frame} com 26 linhas e 8 variaveis.
\describe{
   \item{pac}{identificador do paciente}
   \item{tempo}{tempo em dias ate a morte do paciente}
   \item{trat}{tipo de tratamento recebido}
   \item{idade}{idade do paciente}
   \item{res}{tipo de residuo da doenca (1 se completamente removida, 2 se parcialmente removida)}
   \item{status}{condicao do doente no inicio do estudo (1 se boa, 2 se ruim)}
}}
\source{
Colosimo, E. A.; Giolo, S. R. (2006). Analise de Sobrevivencia Aplicada. Sao Paulo: Edgard Blucher.
}
\usage{
data(ovario)
}
\description{
Dados com efeito de dois tratamentos pos-sirurgicos em pacientes com cancer de ovario.
}
\examples{
data(ovario)


}
\keyword{datasets}
