% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{tm}
\alias{tm}
\title{Mean lifetime
Estimative of Kaplan-Meier survival mean lifetime (with standard error)}
\usage{
tm(fit)
}
\arguments{
\item{fit}{survfit object.}
}
\value{
numeric vector with mean lifetime and standard error
}
\description{
Mean lifetime
Estimative of Kaplan-Meier survival mean lifetime (with standard error)
}
\examples{
data('leucemia')
fitl = survfit(Surv(tempos, cens) ~ 1, data = leucemia)
tm(fitl)
}
\author{
Pedro Araujo
}
