% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{tp}
\alias{tp}
\title{Survival function percentile
Estimative of Kaplan-Meyer survival function percentile by linear interpolation}
\usage{
tp(fit, p = 0.5)
}
\arguments{
\item{fit}{survfit object.}

\item{p}{numeric vector with percentiles.}
}
\value{
vector with aproximation of the time.
}
\description{
Survival function percentile
Estimative of Kaplan-Meyer survival function percentile by linear interpolation
}
\examples{
data('leucemia')
fitl = survfit(Surv(tempos, cens) ~ 1, data = leucemia)
tp(fitl, p = c(0.5, 0.6))
}
\author{
Pedro Araujo
}
